reference_model/requirements.txt
commonroad-vehicle-models==2.0.0
cycler==0.11.0
fonttools==4.32.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
scipy==1.8.0
six==1.16.0
